+++ admin/partials/mailarchiver-admin-settings-archiver-edit.php
@@ -23,7 +23,7 @@
 				'handler' => $current_archiver['handler'],
 				'uuid'    => $current_archiver['uuid'],
 			),
-			admin_url( 'options-general.php' )
+			admin_url( 'admin.php' )
 		)
 	);
 	?>